+++ schemaorggo/aggregaterating.go
@@ -10,15 +10,16 @@ type AggregateRating struct {
 
 	// ItemReviewed see : https://schema.org/itemReviewed
 	// The item that is being reviewed/rated.
-	ItemReviewed *Thing `json:"itemReviewed,omitempty"`
+	ItemReviewed *Thing `json:"itemReviewed,omitempty"` // types : Thing
 
 	// RatingCount see : https://schema.org/ratingCount
 	// The count of total number of ratings.
-	RatingCount int `json:"ratingCount,omitempty"`
+	RatingCount float64 `json:"ratingCount,omitempty"` // types : Integer
 
 	// ReviewCount see : https://schema.org/reviewCount
 	// The count of total number of reviews.
-	ReviewCount int `json:"reviewCount,omitempty"`
+	ReviewCount float64 `json:"reviewCount,omitempty"` // types : Integer
+
 }
 
 func (v AggregateRating) MarshalJSONWithTypeContext() ([]byte, error) {
