+++ schemaorggo/bodyofwater.go
@@ -12,7 +12,7 @@ type BodyOfWater struct {
 	// A property-value pair representing an additional characteristics of the entitity, e.g. a product feature or another characteristic for which there is no matching property in schema.org.
 	//
 	// Note: Publishers should be aware that applications designed to use specific schema.org properties (e.g. http://schema.org/width, http://schema.org/color, http://schema.org/gtin13, ...) will typically expect such data to be provided using those properties, rather than using the generic property/value mechanism.
-	AdditionalProperty *PropertyValue `json:"additionalProperty,omitempty"`
+	AdditionalProperty *PropertyValue `json:"additionalProperty,omitempty"` // types : PropertyValue
 
 	// Address see : https://schema.org/address
 	// Physical address of the item.
@@ -20,33 +20,33 @@ type BodyOfWater struct {
 
 	// AggregateRating see : https://schema.org/aggregateRating
 	// The overall rating, based on a collection of reviews or ratings, of the item.
-	AggregateRating *AggregateRating `json:"aggregateRating,omitempty"`
+	AggregateRating *AggregateRating `json:"aggregateRating,omitempty"` // types : AggregateRating
 
 	// AmenityFeature see : https://schema.org/amenityFeature
 	// An amenity feature (e.g. a characteristic or service) of the Accommodation. This generic property does not make a statement about whether the feature is included in an offer for the main accommodation or available at extra costs.
-	AmenityFeature *LocationFeatureSpecification `json:"amenityFeature,omitempty"`
+	AmenityFeature *LocationFeatureSpecification `json:"amenityFeature,omitempty"` // types : LocationFeatureSpecification
 
 	// BranchCode see : https://schema.org/branchCode
 	// A short textual code (also called "store code") that uniquely identifies a place of business. The code is typically assigned by the parentOrganization and used in structured URLs.
 	//
 	// For example, in the URL http://www.starbucks.co.uk/store-locator/etc/detail/3047 the code "3047" is a branchCode for a particular branch.
-	BranchCode string `json:"branchCode,omitempty"`
+	BranchCode string `json:"branchCode,omitempty"` // types : Text
 
 	// ContainedInPlace see : https://schema.org/containedInPlace
 	// The basic containment relation between a place and one that contains it. Supersedes containedIn (see: https://schema.org/containedIn). Inverse property: containsPlace (see: https://schema.org/containsPlace).
-	ContainedInPlace *Place `json:"containedInPlace,omitempty"`
+	ContainedInPlace *Place `json:"containedInPlace,omitempty"` // types : Place
 
 	// ContainsPlace see : https://schema.org/containsPlace
 	// The basic containment relation between a place and another that it contains. Inverse property: containedInPlace (see: https://schema.org/containedInPlace).
-	ContainsPlace *Place `json:"containsPlace,omitempty"`
+	ContainsPlace *Place `json:"containsPlace,omitempty"` // types : Place
 
 	// Event see : https://schema.org/event
 	// Upcoming or past event associated with this place, organization, or action. Supersedes events (see: https://schema.org/events).
-	Event *Event `json:"event,omitempty"`
+	Event *Event `json:"event,omitempty"` // types : Event
 
 	// FaxNumber see : https://schema.org/faxNumber
 	// The fax number.
-	FaxNumber string `json:"faxNumber,omitempty"`
+	FaxNumber string `json:"faxNumber,omitempty"` // types : Text
 
 	// Geo see : https://schema.org/geo
 	// The geo coordinates of the place.
@@ -94,7 +94,7 @@ type BodyOfWater struct {
 
 	// GlobalLocationNumber see : https://schema.org/globalLocationNumber
 	// The Global Location Number (see: https://schema.orghttp://www.gs1.org/gln) (GLN, sometimes also referred to as International Location Number or ILN) of the respective organization, person, or place. The GLN is a 13-digit number used to identify parties and physical locations.
-	GlobalLocationNumber string `json:"globalLocationNumber,omitempty"`
+	GlobalLocationNumber string `json:"globalLocationNumber,omitempty"` // types : Text
 
 	// HasMap see : https://schema.org/hasMap
 	// A URL to a map of the place. Supersedes map (see: https://schema.org/map), maps (see: https://schema.org/maps).
@@ -102,11 +102,11 @@ type BodyOfWater struct {
 
 	// IsAccessibleForFree see : https://schema.org/isAccessibleForFree
 	// A flag to signal that the item, event, or place is accessible for free. Supersedes free (see: https://schema.org/free).
-	IsAccessibleForFree bool `json:"isAccessibleForFree,omitempty"`
+	IsAccessibleForFree bool `json:"isAccessibleForFree,omitempty"` // types : Boolean
 
 	// IsicV4 see : https://schema.org/isicV4
 	// The International Standard of Industrial Classification of All Economic Activities (ISIC), Revision 4 code for a particular organization, business person, or place.
-	IsicV4 string `json:"isicV4,omitempty"`
+	IsicV4 string `json:"isicV4,omitempty"` // types : Text
 
 	// Logo see : https://schema.org/logo
 	// An associated logo.
@@ -114,11 +114,11 @@ type BodyOfWater struct {
 
 	// MaximumAttendeeCapacity see : https://schema.org/maximumAttendeeCapacity
 	// The total number of individuals that may attend an event or venue.
-	MaximumAttendeeCapacity int `json:"maximumAttendeeCapacity,omitempty"`
+	MaximumAttendeeCapacity float64 `json:"maximumAttendeeCapacity,omitempty"` // types : Integer
 
 	// OpeningHoursSpecification see : https://schema.org/openingHoursSpecification
 	// The opening hours of a certain place.
-	OpeningHoursSpecification *OpeningHoursSpecification `json:"openingHoursSpecification,omitempty"`
+	OpeningHoursSpecification *OpeningHoursSpecification `json:"openingHoursSpecification,omitempty"` // types : OpeningHoursSpecification
 
 	// Photo see : https://schema.org/photo
 	// A photograph of this place. Supersedes photos (see: https://schema.org/photos).
@@ -126,25 +126,26 @@ type BodyOfWater struct {
 
 	// PublicAccess see : https://schema.org/publicAccess
 	// A flag to signal that the Place (see: https://schema.org/Place) is open to public visitors.  If this property is omitted there is no assumed default boolean value
-	PublicAccess bool `json:"publicAccess,omitempty"`
+	PublicAccess bool `json:"publicAccess,omitempty"` // types : Boolean
 
 	// Review see : https://schema.org/review
 	// A review of the item. Supersedes reviews (see: https://schema.org/reviews).
-	Review *Review `json:"review,omitempty"`
+	Review *Review `json:"review,omitempty"` // types : Review
 
 	// SmokingAllowed see : https://schema.org/smokingAllowed
 	// Indicates whether it is allowed to smoke in the place, e.g. in the restaurant, hotel or hotel room.
-	SmokingAllowed bool `json:"smokingAllowed,omitempty"`
+	SmokingAllowed bool `json:"smokingAllowed,omitempty"` // types : Boolean
 
 	// SpecialOpeningHoursSpecification see : https://schema.org/specialOpeningHoursSpecification
 	// The special opening hours of a certain place.
 	//
 	// Use this to explicitly override general opening hours brought in scope by openingHoursSpecification (see: https://schema.org/openingHoursSpecification) or openingHours (see: https://schema.org/openingHours).
-	SpecialOpeningHoursSpecification *OpeningHoursSpecification `json:"specialOpeningHoursSpecification,omitempty"`
+	SpecialOpeningHoursSpecification *OpeningHoursSpecification `json:"specialOpeningHoursSpecification,omitempty"` // types : OpeningHoursSpecification
 
 	// Telephone see : https://schema.org/telephone
 	// The telephone number.
-	Telephone string `json:"telephone,omitempty"`
+	Telephone string `json:"telephone,omitempty"` // types : Text
+
 }
 
 func (v BodyOfWater) MarshalJSONWithTypeContext() ([]byte, error) {
