+++ schemaorggo/creativeworkseason.go
@@ -10,11 +10,11 @@ type CreativeWorkSeason struct {
 
 	// Actor see : https://schema.org/actor
 	// An actor, e.g. in tv, radio, movie, video games etc., or in an event. Actors can be associated with individual items or with a series, episode, clip. Supersedes actors (see: https://schema.org/actors).
-	Actor *Person `json:"actor,omitempty"`
+	Actor *Person `json:"actor,omitempty"` // types : Person
 
 	// Director see : https://schema.org/director
 	// A director of e.g. tv, radio, movie, video gaming etc. content, or of an event. Directors can be associated with individual items or with a series, episode, clip. Supersedes directors (see: https://schema.org/directors).
-	Director *Person `json:"director,omitempty"`
+	Director *Person `json:"director,omitempty"` // types : Person
 
 	// EndDate see : https://schema.org/endDate
 	// The end date and time of the item (in ISO 8601 date format (see: https://schema.orghttp://en.wikipedia.org/wiki/ISO_8601)).
@@ -22,19 +22,19 @@ type CreativeWorkSeason struct {
 
 	// Episode see : https://schema.org/episode
 	// An episode of a tv, radio or game media within a series or season. Supersedes episodes (see: https://schema.org/episodes).
-	Episode *Episode `json:"episode,omitempty"`
+	Episode *Episode `json:"episode,omitempty"` // types : Episode
 
 	// NumberOfEpisodes see : https://schema.org/numberOfEpisodes
 	// The number of episodes in this season or series.
-	NumberOfEpisodes int `json:"numberOfEpisodes,omitempty"`
+	NumberOfEpisodes float64 `json:"numberOfEpisodes,omitempty"` // types : Integer
 
 	// PartOfSeries see : https://schema.org/partOfSeries
 	// The series to which this episode or season belongs. Supersedes partOfTVSeries (see: https://schema.org/partOfTVSeries).
-	PartOfSeries *CreativeWorkSeries `json:"partOfSeries,omitempty"`
+	PartOfSeries *CreativeWorkSeries `json:"partOfSeries,omitempty"` // types : CreativeWorkSeries
 
 	// ProductionCompany see : https://schema.org/productionCompany
 	// The production company or studio responsible for the item e.g. series, video game, episode etc.
-	ProductionCompany *Organization `json:"productionCompany,omitempty"`
+	ProductionCompany *Organization `json:"productionCompany,omitempty"` // types : Organization
 
 	// SeasonNumber see : https://schema.org/seasonNumber
 	// Position of the season within an ordered group of seasons.
@@ -46,7 +46,8 @@ type CreativeWorkSeason struct {
 
 	// Trailer see : https://schema.org/trailer
 	// The trailer of a movie or tv/radio series, season, episode, etc.
-	Trailer *VideoObject `json:"trailer,omitempty"`
+	Trailer *VideoObject `json:"trailer,omitempty"` // types : VideoObject
+
 }
 
 func (v CreativeWorkSeason) MarshalJSONWithTypeContext() ([]byte, error) {
