+++ schemaorggo/scholarlyarticle.go
@@ -10,11 +10,11 @@ type ScholarlyArticle struct {
 
 	// ArticleBody see : https://schema.org/articleBody
 	// The actual body of the article.
-	ArticleBody string `json:"articleBody,omitempty"`
+	ArticleBody string `json:"articleBody,omitempty"` // types : Text
 
 	// ArticleSection see : https://schema.org/articleSection
 	// Articles may belong to one or more 'sections' in a magazine or newspaper, such as Sports, Lifestyle, etc.
-	ArticleSection string `json:"articleSection,omitempty"`
+	ArticleSection string `json:"articleSection,omitempty"` // types : Text
 
 	// PageEnd see : https://schema.org/pageEnd
 	// The page on which the work ends; for example "138" or "xvi".
@@ -26,7 +26,7 @@ type ScholarlyArticle struct {
 
 	// Pagination see : https://schema.org/pagination
 	// Any description of pages that is not separated into pageStart and pageEnd; for example, "1-6, 9, 55" or "10-12, 46-49".
-	Pagination string `json:"pagination,omitempty"`
+	Pagination string `json:"pagination,omitempty"` // types : Text
 
 	// Speakable see : http://pending.schema.org/speakable
 	// Indicates sections of a Web page that are particularly 'speakable' in the sense of being highlighted as being especially appropriate for text-to-speech conversion. Other sections of a page may also be usefully spoken in particular circumstances; the 'speakable' property serves to indicate the parts most likely to be generally useful for speech.
@@ -45,7 +45,8 @@ type ScholarlyArticle struct {
 
 	// WordCount see : https://schema.org/wordCount
 	// The number of words in the text of the Article.
-	WordCount int `json:"wordCount,omitempty"`
+	WordCount float64 `json:"wordCount,omitempty"` // types : Integer
+
 }
 
 func (v ScholarlyArticle) MarshalJSONWithTypeContext() ([]byte, error) {
